pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "commit-and-tag-version"
version = "0.0.1"
description = "A utility for versioning using semver and CHANGELOG generation powered by Conventional Commits"
readme = "README.md"
license = "ISC"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "semver>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.0",
]

[project.scripts]
commit-and-tag-version = "commit_and_tag_version.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
